requirements.txt
python-dotenv==1.2.1
websockets==12.0
requests==2.32.5
structlog==24.1.0
coinbase-advanced-py
hyperliquid-python-sdk
eth-account==0.11.3
